requirements.txt
appdirs==1.4.4
asgiref==3.2.10
autopep8==1.5.6
distlib==0.3.1
Django==3.1.1
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
filelock==3.0.12
gunicorn==20.1.0
mysqlclient==2.0.3
Pillow==8.2.0
pycodestyle==2.7.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.3
whitenoise==5.2.0